+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_image_operation.dart
@@ -71,10 +71,34 @@ abstract class MPTH2FileEditStateImageOperation extends MPTH2FileEditState {
 
   @override
   void onKeyDownEvent(KeyDownEvent event) {
-    if (event.logicalKey == LogicalKeyboardKey.escape) {
-      th2FileEditController.stateController.setState(
-        MPTH2FileEditStateType.selectEmptySelection,
-      );
+    final bool isAltPressed = MPInteractionAux.isAltPressed();
+    final bool isCtrlPressed = MPInteractionAux.isCtrlPressed();
+    final bool isMetaPressed = MPInteractionAux.isMetaPressed();
+    final bool isShiftPressed = MPInteractionAux.isShiftPressed();
+
+    if (isAltPressed || isCtrlPressed || isMetaPressed || isShiftPressed) {
+      if (event.logicalKey == LogicalKeyboardKey.escape) {
+        th2FileEditController.stateController.setState(
+          MPTH2FileEditStateType.selectEmptySelection,
+        );
+      }
+
+      return;
+    }
+
+    switch (event.logicalKey) {
+      case LogicalKeyboardKey.escape:
+        th2FileEditController.stateController.setState(
+          MPTH2FileEditStateType.selectEmptySelection,
+        );
+      case LogicalKeyboardKey.keyH:
+        th2FileEditController.stateController.onButtonPressed(
+          MPButtonType.flipImageHorizontally,
+        );
+      case LogicalKeyboardKey.keyV:
+        th2FileEditController.stateController.onButtonPressed(
+          MPButtonType.flipImageVertically,
+        );
     }
   }
 }