+++ lib/src/state_machine/mp_th2_file_edit_state_machine/types/mp_button_type.dart
@@ -17,6 +17,8 @@ enum MPButtonType {
   createMapConnectionLines,
   cutElements,
   duplicateElements,
+  flipImageHorizontally,
+  flipImageVertically,
   joinLinesAtCoincidingExtremities,
   mergeAreas,
   nodeEdit,